+++ src/bolos/io/sdk/include/errors.h
@@ -19,6 +19,8 @@
 #if !defined(ERRORS_H)
 #define ERRORS_H
 
+#include "os_errors.h"
+
 /**
  * Applications-reserved error codes ranges.
  * The Operating System do not use any error code within these ranges.
@@ -123,8 +125,6 @@ enum sdk_generic_identifiers {
 #define SWO_IOL_BLE_0B (ERR_IOL_BLE + ERR_GEN_ID_0B) // 0x150B
 #define SWO_IOL_BLE_0C (ERR_IOL_BLE + ERR_GEN_ID_0C) // 0x150C
 #endif                                               // HAVE_BLE
-
-#define SWO_SEC_PIN_15 0x5515
 
 /**
  * The process is successful.