+++ src/bolos/os_bip32.c
@@ -5,15 +5,25 @@
 
 #include "os_bip32.h"
 
+#include "appflags.h"
 #include "bolos/exception.h"
 #include "cx.h"
 #include "cx_ec.h"
 #include "cx_utils.h"
 #include "emulate.h"
 #include "environment.h"
+#include "launcher.h"
 
 #define BIP32_HARDEN_MASK      0x80000000
 #define BIP32_SECP_SEED_LENGTH 12
+
+#define DERIVE_AUTH_256K1      0x01
+#define DERIVE_AUTH_256R1      0x02
+#define DERIVE_AUTH_ED25519    0x04
+#define DERIVE_AUTH_SLIP21     0x08
+#define DERIVE_AUTH_BLS12381G1 0x10
+
+#define BIP32_WILDCARD_VALUE (0x7fffffff)
 
 #define cx_ecfp_generate_pair     sys_cx_ecfp_generate_pair
 #define cx_ecfp_init_private_key  sys_cx_ecfp_init_private_key
@@ -31,6 +41,11 @@ static uint8_t const BIP32_ED_SEED[] = { 'e', 'd', '2', '5', '5', '1',
 static uint8_t const SLIP21_SEED[] = { 'S', 'y', 'm', 'm', 'e', 't',
                                        'r', 'i', 'c', ' ', 'k', 'e',
                                        'y', ' ', 's', 'e', 'e', 'd' };
+
+static uint32_t sys_os_perso_derive_node_with_seed_key_internal(
+    uint32_t mode, cx_curve_t curve, const uint32_t *path, uint32_t pathLength,
+    uint8_t *privateKey, uint8_t *chain, uint8_t *seed_key,
+    uint32_t seed_key_length, bool fromApp);
 
 static bool is_hardened_child(uint32_t child)
 {
@@ -71,6 +86,94 @@ static ssize_t get_seed_key_slip21(const uint8_t **sk)
   sk_length = sizeof(SLIP21_SEED);
 
   return sk_length;
+}
+
+static void os_perso_derive_node_bip32_check_path(unsigned int mode,
+                                                  cx_curve_t curve,
+                                                  const unsigned int *path,
+                                                  unsigned int pathLength,
+                                                  bool fromApp)
+{
+  unsigned char pathValid = 0;
+  unsigned int derive_path_length;
+  unsigned char *derive_path;
+
+  // TODO : Harden against glitches
+  // Check the path against the current application path
+  derive_path_length = (fromApp) ? get_app_derivation_path(&derive_path) : 0;
+
+  // No information specified, everything is valid
+  if (derive_path_length != 0) {
+    unsigned char curveMask;
+    unsigned int checkOffset = 1;
+    // Abort if not authorized to operate on this curve
+    if (mode == HDW_SLIP21) {
+      curveMask = DERIVE_AUTH_SLIP21;
+    } else {
+      switch (curve) {
+      case CX_CURVE_256K1:
+        curveMask = DERIVE_AUTH_256K1;
+        break;
+      case CX_CURVE_256R1:
+        curveMask = DERIVE_AUTH_256R1;
+        break;
+      case CX_CURVE_Ed25519:
+        curveMask = DERIVE_AUTH_ED25519;
+        break;
+      default:
+        THROW(SWO_PAR_VAL_13);
+      }
+    }
+    if ((derive_path[0] & curveMask) != curveMask) {
+      THROW(SWO_PAR_VAL_14);
+    }
+    // if only the curve was specified, all paths are valid
+    if (derive_path_length > 1) {
+      while ((checkOffset < derive_path_length) && !pathValid) {
+        unsigned char subPathLength = derive_path[checkOffset];
+        unsigned char slip21Marker = ((subPathLength & 0x80) != 0);
+        unsigned char slip21Request = (mode == HDW_SLIP21);
+        subPathLength &= 0x7F;
+        // Authorize if a subpath is valid
+        if (pathLength >= subPathLength) {
+          pathValid = 1;
+          if (slip21Marker != slip21Request) {
+            pathValid = 0;
+          } else {
+            if (slip21Request) {
+              // compare path as bytes as slip21 path may be a string.
+              // NOTE: interpret path first byte (len) as a number of bytes in
+              // the path
+              if (memcmp(path, &derive_path[checkOffset + 1], subPathLength)) {
+                pathValid = 0;
+              }
+            } else {
+              for (unsigned int i = 0; i < subPathLength; i++) {
+                uint32_t allowed_prefix =
+                    U4BE(derive_path, (checkOffset + 1 + 4 * i));
+                if (path[i] != allowed_prefix) {
+                  // path is considered to be valid if the allowed prefix
+                  // contains the 'wildcard' value 0x7fffffff
+                  if (allowed_prefix != BIP32_WILDCARD_VALUE) {
+                    pathValid = 0;
+                    break;
+                  }
+                }
+              }
+            }
+          }
+        }
+        checkOffset += 1 + (slip21Marker ? 1 : 4) * subPathLength;
+      }
+    } else {
+      pathValid = 1;
+    }
+  } else {
+    pathValid = 1;
+  }
+  if (!pathValid) {
+    THROW(SWO_PAR_VAL_15);
+  }
 }
 
 void expand_seed_ed25519(const uint8_t *sk, size_t sk_length, uint8_t *seed,
@@ -404,6 +507,16 @@ unsigned long sys_os_perso_derive_node_with_seed_key(
     unsigned int pathLength, unsigned char *privateKey, unsigned char *chain,
     unsigned char *seed_key, unsigned int seed_key_length)
 {
+  return sys_os_perso_derive_node_with_seed_key_internal(
+      mode, curve, path, pathLength, privateKey, chain, seed_key,
+      seed_key_length, true);
+}
+
+static uint32_t sys_os_perso_derive_node_with_seed_key_internal(
+    uint32_t mode, cx_curve_t curve, const uint32_t *path, uint32_t pathLength,
+    uint8_t *privateKey, uint8_t *chain, uint8_t *seed_key,
+    uint32_t seed_key_length, bool fromApp)
+{
   ssize_t sk_length;
   size_t seed_size;
   uint8_t seed[MAX_SEED_SIZE];
@@ -430,6 +543,23 @@ unsigned long sys_os_perso_derive_node_with_seed_key(
     curve = CX_CURVE_Curve448;
     break;
   }
+
+  // If APPLICATION_FLAG_DERIVE_MASTER is not set for an application, forbid
+  // deriving from seed with a null path
+  if ((mode != HDW_SLIP21) && (fromApp) &&
+      !(app_flags & APPLICATION_FLAG_DERIVE_MASTER)) {
+    uint32_t i;
+    for (i = 0; i < pathLength; i++) {
+      if ((path[i] & 0x80000000) != 0) {
+        break;
+      }
+    }
+    if (i == pathLength) {
+      THROW(SWO_PAR_VAL_12);
+    }
+  }
+
+  os_perso_derive_node_bip32_check_path(mode, curve, path, pathLength, fromApp);
 
   if (seed_key == NULL || seed_key_length == 0) {
     if (mode != HDW_SLIP21) {
@@ -504,10 +634,11 @@ unsigned long sys_os_perso_get_master_key_identifier(uint8_t *identifier,
   unsigned int path = 0;
 
   /* Getting raw private key */
-  unsigned long ret = sys_os_perso_derive_node_with_seed_key(
-      HDW_NORMAL, CX_CURVE_SECP256K1, &path, 0, raw_private_key, NULL, NULL, 0);
+  uint32_t ret = sys_os_perso_derive_node_with_seed_key_internal(
+      HDW_NORMAL, CX_CURVE_SECP256K1, &path, 0, raw_private_key, NULL, NULL, 0,
+      false);
   if (ret != 0) {
-    errx(1, "%s: sys_os_perso_derive_node_with_seed_key() failed with %lu",
+    errx(1, "%s: sys_os_perso_derive_node_with_seed_key() failed with %u",
          __func__, ret);
     THROW(EXCEPTION);
   }