+++ CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.25.10] 2025-11-20
+## [0.25.10] 2025-12-04
 
 ### Added
 - Support throwing exceptions from OS (mainly Cx syscalls) like on real devices
+- Support retrieving appflags from elf and use it for derivation
+- Support retrieving derivation_path from elf and use it for derivation
 
 ## [0.25.9] 2025-11-19